Factor out StdioSetter and Stdio.WithInitial().
